--- src/components/Hello.js
@@ -1,8 +1,0 @@
-import React from 'react';
-
-const Hello = (props) =>{
-  console.log(props.match)
-  return <div> hello </div>
-}
-
-export default Hello
