--- src/components/Welcome.js
@@ -1,7 +1,0 @@
-import React  from 'react';
-
-const Welcome = (props) =>{
-  return <div> welcome</div>
-}
-
-export default Welcome
