--- src/components/index.js
@@ -1,5 +1,0 @@
-
-export {default as Hello} from './Hello.js'
-export {default as Welcome} from './Welcome'
-export {default as Pokemon} from './Pokemon'
-export {default as SinglePokemon} from './SinglePokemon'
