+++ src/index.js
@@ -1,12 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import './index.css';
 import App from './App';
-import {BrowserRouter as Router} from 'react-router-dom'
 
-
-ReactDOM.render(
-<Router>
-  <App/>
- </Router>, 
- document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
